Initial attempt to wire basic socket functionality up
Now that the TCP socket handling is ready, test it with basic HTTP
response. It receives data from the client, but currently doesn't parse
anything. It always responds with HTTP 200 OK.

This is a good starting point. It's a validation that the socket library
I develop is now working properly.

Next, I'll try to use an external library, picohttpparser, to parse HTTP
requests.

Signed-off-by: Alviro Iskandar Setiawan <alviro.iskandar@gnuweeb.org>