main: Add signal handler for clean shutdown
Now that ais_sock_tcp_srv_stop() can be used to shut the server down
gracefully. Use it via signal handler. So when we hit CTRL + C, the exit
process will be clean, the connection will be closed one by one as well.
It's very good to handle them cleanly.

Signed-off-by: Alviro Iskandar Setiawan <alviro.iskandar@gnuweeb.org>