tcp: Fix a use-after-free bug during shutdown
The srv->clients slot was not set to NULL after the client struct gets
freed. Add a new field, idx, in the client struct to keep track where
the current client is located in the client slot.

Make sure when closing a connection, the unused slot doesn't contain a
dangling pointer. Swap it with the last index if the freed index is not
the last index.

Fixes: 9ea19ad700e8e0caf19d45ebb995d3a1549d8b90 ("Initial commit")
Signed-off-by: Alviro Iskandar Setiawan <alviro.iskandar@gnuweeb.org>